requirements.txt
blessed
click
arrow
gitpython
